+++ examples/monaco-graphql-webpack/src/index.ts
@@ -4,9 +4,9 @@ import 'regenerator-runtime/runtime';
 import 'monaco-graphql/esm/monaco.contribution';
 
 // eslint-disable-next-line spaced-comment
-/// <reference path='../../../node_modules/monaco-editor/monaco.d.ts'/>
+/// <reference types='monaco-editor'/>
 // eslint-disable-next-line spaced-comment
-/// <reference path='../../../packages/monaco-graphql/src/typings/monaco.d.ts'/>
+///  <reference types='monaco-graphql'/>
 
 // NOTE: using loader syntax becuase Yaml worker imports editor.worker directly and that
 // import shouldn't go through loader syntax.
@@ -99,7 +99,17 @@ const operationEditor = monaco.editor.create(
   },
 );
 
-monaco.languages.graphql.graphqlDefaults.setSchemaConfig({ uri: SCHEMA_URL });
+monaco.languages.graphql.graphqlDefaults.setFormattingOptions({
+  // TODO: get these types working properly
+  prettierConfig: {
+    printWith: 120,
+  },
+});
+
+monaco.languages.graphql.graphqlDefaults.setSchemaConfig({
+  // TODO: get these types working properly
+  uri: SCHEMA_URL,
+});
 
 /**
  * Basic Operation Exec Example