+++ packages/graphiql/src/components/QueryHistory.tsx
@@ -21,6 +21,7 @@ const MAX_HISTORY_LENGTH = 20;
 const shouldSaveQuery = (
   query?: string,
   variables?: string,
+  headers?: string,
   lastQuerySaved?: QueryStoreItem,
 ) => {
   if (!query) {
@@ -44,7 +45,12 @@ const shouldSaveQuery = (
     if (
       JSON.stringify(variables) === JSON.stringify(lastQuerySaved.variables)
     ) {
-      return false;
+      if (JSON.stringify(headers) === JSON.stringify(lastQuerySaved.headers)) {
+        return false;
+      }
+      if (headers && !lastQuerySaved.headers) {
+        return false;
+      }
     }
     if (variables && !lastQuerySaved.variables) {
       return false;
@@ -56,6 +62,7 @@ const shouldSaveQuery = (
 type QueryHistoryProps = {
   query?: string;
   variables?: string;
+  headers?: string;
   operationName?: string;
   queryID?: number;
   onSelectQuery: HandleSelectQueryFn;
@@ -116,12 +123,21 @@ export class QueryHistory extends React.Component<
   updateHistory = (
     query?: string,
     variables?: string,
+    headers?: string,
     operationName?: string,
   ) => {
-    if (shouldSaveQuery(query, variables, this.historyStore.fetchRecent())) {
+    if (
+      shouldSaveQuery(
+        query,
+        variables,
+        headers,
+        this.historyStore.fetchRecent(),
+      )
+    ) {
       this.historyStore.push({
         query,
         variables,
+        headers,
         operationName,
       });
       const historyQueries = this.historyStore.items;
@@ -137,6 +153,7 @@ export class QueryHistory extends React.Component<
   toggleFavorite: HandleToggleFavoriteFn = (
     query,
     variables,
+    headers,
     operationName,
     label,
     favorite,
@@ -144,6 +161,7 @@ export class QueryHistory extends React.Component<
     const item: QueryStoreItem = {
       query,
       variables,
+      headers,
       operationName,
       label,
     };
@@ -163,6 +181,7 @@ export class QueryHistory extends React.Component<
   editLabel: HandleEditLabelFn = (
     query,
     variables,
+    headers,
     operationName,
     label,
     favorite,
@@ -170,6 +189,7 @@ export class QueryHistory extends React.Component<
     const item = {
       query,
       variables,
+      headers,
       operationName,
       label,
     };