+++ packages/graphiql/src/components/__tests__/GraphiQL.spec.tsx
@@ -18,6 +18,8 @@ import {
   mockBadQuery,
   mockQuery2,
   mockVariables2,
+  mockHeaders1,
+  mockHeaders2,
 } from './fixtures';
 
 codeMirrorModules.forEach(m => jest.mock(m, () => {}));
@@ -139,17 +141,15 @@ describe('GraphiQL', () => {
     const { container: container1 } = render(
       <GraphiQL fetcher={noOpFetcher} />,
     );
-    const queryVariables = container1.querySelector(
-      '[aria-label="Query Variables"]',
-    );
+    const queryVariables = container1.querySelector('.variable-editor');
 
     expect(queryVariables.style.height).toEqual('');
 
-    const variableEditorTitle = container1.querySelector(
-      '#variable-editor-title',
+    const secondaryEditorTitle = container1.querySelector(
+      '#secondary-editor-title',
     );
-    fireEvent.mouseDown(variableEditorTitle);
-    fireEvent.mouseMove(variableEditorTitle);
+    fireEvent.mouseDown(secondaryEditorTitle);
+    fireEvent.mouseMove(secondaryEditorTitle);
     expect(queryVariables.style.height).toEqual('200px');
 
     const { container: container2 } = render(
@@ -166,9 +166,7 @@ describe('GraphiQL', () => {
         defaultVariableEditorOpen={false}
       />,
     );
-    const queryVariables3 = container3.querySelector(
-      '[aria-label="Query Variables"]',
-    );
+    const queryVariables3 = container3.querySelector('.variable-editor');
     expect(queryVariables3?.style.height).toEqual('');
   });
 
@@ -177,6 +175,7 @@ describe('GraphiQL', () => {
       <GraphiQL
         query={mockQuery1}
         variables={mockVariables1}
+        headers={mockHeaders1}
         operationName={mockOperationName1}
         fetcher={noOpFetcher}
       />,
@@ -200,6 +199,7 @@ describe('GraphiQL', () => {
         operationName={mockOperationName1}
         query={mockQuery1}
         variables={mockVariables1}
+        headers={mockHeaders1}
       />,
     );
     fireEvent.click(getByTitle('Execute Query (Ctrl-Enter)'));
@@ -213,6 +213,7 @@ describe('GraphiQL', () => {
         operationName={mockOperationName1}
         query={mockQuery1}
         variables={mockVariables1}
+        headers={mockHeaders1}
       />,
     );
     fireEvent.click(getByTitle('Execute Query (Ctrl-Enter)'));
@@ -228,6 +229,7 @@ describe('GraphiQL', () => {
         operationName={mockOperationName1}
         query={mockQuery1}
         variables={mockVariables1}
+        headers={mockHeaders1}
       />,
     );
     const executeQueryButton = getByTitle('Execute Query (Ctrl-Enter)');
@@ -255,6 +257,7 @@ describe('GraphiQL', () => {
         operationName={mockOperationName1}
         query={mockQuery1}
         variables={mockVariables1}
+        headers={mockHeaders1}
       />,
     );
     const executeQueryButton = getByTitle('Execute Query (Ctrl-Enter)');
@@ -265,6 +268,34 @@ describe('GraphiQL', () => {
       container.querySelector('[aria-label="Query Variables"] .mockCodeMirror'),
       {
         target: { value: mockVariables2 },
+      },
+    );
+
+    fireEvent.click(executeQueryButton);
+    expect(container.querySelectorAll('.history-label')).toHaveLength(2);
+  });
+
+  it('will save query if headers are different ', () => {
+    const { getByTitle, getByText, container } = render(
+      <GraphiQL
+        fetcher={noOpFetcher}
+        operationName={mockOperationName1}
+        query={mockQuery1}
+        variables={mockVariables1}
+        headers={mockHeaders1}
+        headerEditorEnabled
+      />,
+    );
+    const executeQueryButton = getByTitle('Execute Query (Ctrl-Enter)');
+    fireEvent.click(executeQueryButton);
+    expect(container.querySelectorAll('.history-label')).toHaveLength(1);
+
+    fireEvent.click(getByText('Request Headers'));
+
+    fireEvent.change(
+      container.querySelector('[aria-label="Request Headers"] .mockCodeMirror'),
+      {
+        target: { value: mockHeaders2 },
       },
     );